Changes to the header to add the Help button and information
Including only hiding the info and showing the button if javascript is running.

Still need to remove transition from the initial hide on page load.

Also need to improve how it looks on small screen when signed in.
