+++ common/assets/js/app/modules/help.js
@@ -1,30 +1,31 @@
 import domready from './domready'
 
-domready(() => {
-  const btn = document.querySelector('.js-help-btn')
-  const help = document.querySelector('.js-help-body')
-  const classClosed = 'is-closed'
 
-  let openedByClick = false
+domready(() => {
+  const btn = document.querySelector('.js-help-btn');
+  const help = document.querySelector('.js-help-body');
+  const classClosed = 'is-closed';
+
+  let openedByClick = false;
 
   if (help === null || btn === null) {
-    return false
+    return false;
   }
 
+
   help.addEventListener('focus', e => {
-    help.classList.remove(classClosed)
-  }, true)
+    help.classList.remove(classClosed);
+  }, true);
 
   help.addEventListener('blur', e => {
     if (!openedByClick) {
-      help.classList.add(classClosed)
+      help.classList.add(classClosed);
     }
-  }, true)
+  }, true);
 
-  btn.classList.remove('u-hidden')
   btn.addEventListener('click', e => {
-    e.preventDefault()
-    openedByClick = !openedByClick
-    help.classList.toggle(classClosed)
-  })
+    e.preventDefault();
+    openedByClick = !openedByClick;
+    help.classList.toggle(classClosed);
+  });
 })
