+++ frontstage/assets/js/app/main.js
@@ -6,6 +6,7 @@ import domready from '../../../../common/assets/js/app/modules/domready';
 import '../../../../common/assets/js/app/modules/details-toggle';
 import '../../../../common/assets/js/app/modules/inpagelink';
 import '../../../../common/assets/js/app/modules/focus-styles';
+import './modules/help';
 
 import { default as passwordObfuscation } from './modules/password-obfuscation.dom';
 import { default as passwordValidation } from './modules/password-validation.dom';
