+++ frontstage/static/css/default/fixed.css
@@ -4648,6 +4648,10 @@ input[type=file]::-webkit-file-upload-button:hover {
   max-height: 40rem;
 }
 
+.has-js .header__help.is-closed--init {
+  max-height: 0;
+}
+
 .has-js .header__help.is-closed {
   max-height: 0;
   -webkit-transition: max-height 200ms ease-out;
