+++ frontstage/static/css/default/responsive.css
@@ -6743,6 +6743,10 @@ input[type=file]::-webkit-file-upload-button:hover {
   max-height: 40rem;
 }
 
+.has-js .header__help.is-closed--init {
+  max-height: 0;
+}
+
 .has-js .header__help.is-closed {
   max-height: 0;
   -webkit-transition: max-height 200ms ease-out;
