+++ frontstage/templates/layouts/_base.html
@@ -83,38 +83,6 @@
     {% endblock scripts %}
   	<!--<![endif]-->
 
-    <script>
-    $(function () {
-      var btn = document.querySelector('.js-help-btn');
-      var help = document.querySelector('.js-help-body');
-      var classClosed = 'is-closed';
-
-      var openedByClick = false;
-
-      if (help === null || btn === null) {
-        return false;
-      }
-
-      help.addEventListener('focus', function (e) {
-        help.classList.remove(classClosed);
-      }, true);
-
-      help.addEventListener('blur', function (e) {
-        if (!openedByClick) {
-          help.classList.add(classClosed);
-        }
-      }, true);
-
-      btn.classList.remove('u-hidden');
-      btn.addEventListener('click', function (e) {
-        e.preventDefault();
-        openedByClick = !openedByClick;
-        help.classList.toggle(classClosed);
-      });
-    });
-    </script>
-
-
   </body>
 
 </html>
