+++ frontstage/templates/partials/header.html
@@ -1,6 +1,6 @@
 
-<div class="helpinfo">
-  <div class="header__help is-closed js-help-body">
+<div id="helpinfo" class="helpinfo">
+  <div class="header__help js-help-body">
     <div class="header__container container">
       <div class="header__title">Help –</div>
       <ul class="header__nav nav nav--horizontal">
@@ -35,7 +35,7 @@
              class="logo__img"/>
     </div>
 
-    <button class="header__btn btn btn--light btn--border js-help-btn">
+    <button aria-controls="helpinfo" class="header__btn btn btn--light btn--border js-help-btn">
       Help
     </button>
